fix: Rename AABB to BoundingBox

+++ GTLEngine/Source/Launch/LaunchWindows.cpp
@@ -55,7 +55,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine
     uint32 Width = 1600;
     uint32 Height = 900;
 
-    RECT rc = { 0, 0, Width, Height };
+    RECT rc = { 0, 0, LONG(Width), LONG(Height) };
     AdjustWindowRect(&rc, WS_OVERLAPPEDWINDOW, FALSE);
     Width = rc.right - rc.left;
     Height = rc.bottom - rc.top;
